LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/foreign-fetch-basics.https-expected.txt

CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/service-workers/service-worker/resources/foreign-fetch/scope/wildcard/foo?basic. Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/service-workers/service-worker/resources/foreign-fetch/scope/wildcard/intercept/foo?basic. Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/service-workers/service-worker/resources/foreign-fetch/scope/match-origin/intercept/foo?basic. Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/service-workers/service-worker/resources/foreign-fetch/scope/nomatch-origin/foo?basic. Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/service-workers/service-worker/resources/foreign-fetch/scope/origin-list/intercept/foo?basic. Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/service-workers/service-worker/resources/foreign-fetch/scope/controlled?basic. Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Cross-origin script load denied by Cross-Origin Resource Sharing policy.
  

Harness Error (TIMEOUT), message = null

FAIL Service Worker intercepts fetches in scope with wildcard origin. promise_test: Unhandled rejection with value: object "TypeError: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin."
FAIL Service Worker intercepts fetches in scope with explicit origin. promise_test: Unhandled rejection with value: object "TypeError: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin."
PASS Service Worker doesn't intercept fetches with non matching origin. 
FAIL Service Worker intercepts fetches in scope with explicit origin list. promise_test: Unhandled rejection with value: object "TypeError: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin."
FAIL Service Worker does not intercept same origin fetches. promise_test: Unhandled rejection with value: object "TypeError: Type error"
PASS Service Worker does not intercept navigations. 
FAIL Service Worker that fallback to network should fallback to network. promise_test: Unhandled rejection with value: object "TypeError: Type error"
FAIL Service Worker that fetch from the network should fallback to network. promise_test: Unhandled rejection with value: object "TypeError: Type error"
FAIL Falling back to network should still respect CORS. promise_test: Unhandled rejection with value: object "TypeError: Type error"
FAIL Foreign fetch can intercept fetch requests from SW controlled pages. promise_test: Unhandled rejection with value: object "TypeError: Origin https://localhost:9443 is not allowed by Access-Control-Allow-Origin."
TIMEOUT Foreign fetch can intercept resource requests from SW controlled pages. Test timed out
NOTRUN Referrer and origin are set correctly in ForeignFetchEvent. 
NOTRUN Service Worker does not intercept fetches from an insecure context. 
NOTRUN Service Worker does not intercept fetches with CORS preflight 
NOTRUN Foreign fetch rejects if resolved with a null response. 
